优化 activate 逻辑
